requirements.txt
appdirs==1.4.3
certifi==2019.9.11
chardet==3.0.4
Click==7.0
distlib==0.3.0
Django==2.2.7
filelock==3.0.12
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.18
protobuf==3.10.0
PyQt5==5.13.2
PyQt5-sip==12.7.1
pyqt5-tools==5.13.2.1.6rc1
python-dotenv==0.12.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
virtualenv==20.0.1
Werkzeug==0.16.0
